+++ CHANGES.md
@@ -14,7 +14,8 @@
 
 ### UPDATE
 
-- Android Studio 3.1.1 に対応した
+- Android Studio 3.1.2 に対応した
+- Kotlin を 1.2.41 に上げた
 
 ### ADD